+++ apps/mobile/src/components/Boundaries/TokenFailureBoundary/TokenFailureBoundary.tsx
@@ -1,136 +1,20 @@
-import { createElement, PropsWithChildren, useCallback, useEffect, useMemo } from 'react';
-import { Text, View } from 'react-native';
+import { createElement, useCallback, useEffect, useMemo } from 'react';
 import { fetchQuery, graphql, useFragment, useRelayEnvironment } from 'react-relay';
-import { ErrorIcon } from 'src/icons/ErrorIcon';
 
 import { useTokenStateManagerContext } from '~/contexts/TokenStateManagerContext';
-import { TokenFailureBoundaryErrorFallbackFragment$key } from '~/generated/TokenFailureBoundaryErrorFallbackFragment.graphql';
 import { TokenFailureBoundaryFragment$key } from '~/generated/TokenFailureBoundaryFragment.graphql';
-import { TokenFailureBoundaryLoadingFallbackFragment$key } from '~/generated/TokenFailureBoundaryLoadingFallbackFragment.graphql';
 import { TokenFailureBoundaryPollerQuery } from '~/generated/TokenFailureBoundaryPollerQuery.graphql';
 import {
   ReportingErrorBoundary,
   ReportingErrorBoundaryProps,
 } from '~/shared/errors/ReportingErrorBoundary';
 
-type FallbackProps = {
-  fallbackAspectSquare?: boolean;
-  variant?: 'tiny' | 'small' | 'medium' | 'large';
-};
-
-function FallbackWrapper({ children, fallbackAspectSquare }: PropsWithChildren<FallbackProps>) {
-  const inner = useMemo(
-    () => (
-      <View className="w-full h-full bg-porcelain dark:bg-black-800 flex items-center justify-center text-center truncate p-1">
-        {children}
-      </View>
-    ),
-    [children]
-  );
-
-  if (fallbackAspectSquare) {
-    return <View className="aspect-square">{inner}</View>;
-  }
-
-  return inner;
-}
-
-function variantToTextSize(variant: FallbackProps['variant'] = 'tiny') {
-  return {
-    tiny: 'xs',
-    small: 'xs',
-    medium: 'sm',
-    large: 'base',
-  }[variant];
-}
-
-function variantToSubtextSize(variant: FallbackProps['variant'] = 'tiny') {
-  return {
-    tiny: 'xxs',
-    small: 'xxs',
-    medium: 'xs',
-    large: 'xs',
-  }[variant];
-}
-
-// TODO v soon:
-// - use refresh icon instead, and fetch new asset
-// - responsiveness for tiny and large versions
-// - check dark mode
-// - truncate / ellipses text
-function TokenPreviewErrorFallback({
-  tokenRef,
-  variant,
-}: // fallbackAspectSquare,
-// variant,
-{
-  tokenRef: TokenFailureBoundaryErrorFallbackFragment$key;
-} & FallbackProps) {
-  const token = useFragment(
-    graphql`
-      fragment TokenFailureBoundaryErrorFallbackFragment on Token {
-        tokenId
-        contract {
-          name
-        }
-      }
-    `,
-    tokenRef
-  );
-
-  return (
-    <>
-      <Text
-        className={`text-${variantToTextSize(variant)} text-metal text-center`}
-        numberOfLines={2}
-      >
-        {token.contract?.name ?? token.tokenId}
-      </Text>
-      {variant === 'tiny' ? null : (
-        <View className="p-1">
-          <ErrorIcon />
-        </View>
-      )}
-    </>
-  );
-}
-
-// TODO v soon:
-// - responsiveness for tiny and large versions
-// - check dark mode
-// - truncate / ellipses text
-function TokenPreviewLoadingFallback({
-  tokenRef,
-  variant,
-}: {
-  tokenRef: TokenFailureBoundaryLoadingFallbackFragment$key;
-} & FallbackProps) {
-  const token = useFragment(
-    graphql`
-      fragment TokenFailureBoundaryLoadingFallbackFragment on Token {
-        tokenId
-        contract {
-          name
-        }
-      }
-    `,
-    tokenRef
-  );
-
-  return (
-    <>
-      <Text
-        className={`text-${variantToTextSize(variant)} text-metal text-center`}
-        numberOfLines={2}
-      >
-        {token.contract?.name ?? token.tokenId}
-      </Text>
-      {variant === 'tiny' ? null : (
-        <Text className={`text-${variantToSubtextSize(variant)} text-metal`}>(processing)</Text>
-      )}
-    </>
-  );
-}
+import {
+  FallbackProps,
+  FallbackWrapper,
+  TokenPreviewErrorFallback,
+  TokenPreviewLoadingFallback,
+} from './TokenFailureFallbacks';
 
 type Props = {
   tokenRef: TokenFailureBoundaryFragment$key;
@@ -151,8 +35,8 @@ export function TokenFailureBoundary({
     graphql`
       fragment TokenFailureBoundaryFragment on Token {
         dbid
-        ...TokenFailureBoundaryErrorFallbackFragment
-        ...TokenFailureBoundaryLoadingFallbackFragment
+        ...TokenFailureFallbacksErrorFallbackFragment
+        ...TokenFailureFallbacksLoadingFallbackFragment
       }
     `,
     tokenRef
