+++ apps/mobile/src/components/NftPreview/NftPreview.tsx
@@ -18,7 +18,7 @@ import {
   useGetSinglePreviewImageProps,
 } from '~/shared/relay/useGetPreviewImages';
 
-import { TokenFailureBoundary } from '../Boundaries/TokenFailureBoundary';
+import { TokenFailureBoundary } from '../Boundaries/TokenFailureBoundary/TokenFailureBoundary';
 import { GallerySkeleton } from '../GallerySkeleton';
 import { ImageState, NftPreviewSharedProps } from './UniversalNftPreview';
